% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{plot.venn}
\alias{plot.venn}
\title{Plot \link{venn} Object}
\usage{
\method{plot}{venn}(x, ...)
}
\arguments{
\item{x}{a \link{venn} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{plot.venn} does not have a returned value.
}
\description{
Plot \link{venn} Object
}
