% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{print.venn}
\alias{print.venn}
\title{Print \link{venn} Object}
\usage{
\method{print}{venn}(x, ...)
}
\arguments{
\item{x}{a \link{venn} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{print.venn} does not have a returned value.
}
\description{
Print \link{venn} Object
}
