% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprintf.R
\name{Sprintf.venn}
\alias{Sprintf.venn}
\title{Sprintf.venn}
\usage{
Sprintf.venn(x)
}
\arguments{
\item{x}{a \link{venn} object}
}
\value{
Function \link{Sprintf.venn} returns a \link[base]{character} scalar.
}
\description{
Sprintf.venn
}
