% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{venn}
\alias{venn}
\alias{venn.list}
\alias{venn.data.frame}
\alias{venn.matrix}
\title{Venn Diagram via \CRANpkg{VennDiagram}}
\usage{
venn(object, ...)

\method{venn}{list}(object, ...)

\method{venn}{data.frame}(object, ...)

\method{venn}{matrix}(
  object,
  palette = c("ggplot", "rainbow"),
  ind = FALSE,
  lty = "blank",
  fill = switch(match.arg(palette), ggplot = pal_hue(), rainbow = rainbow)(n = n_cat),
  alpha = 0.25,
  cex = 1,
  cat.col = fill,
  cat.fontface = "bold",
  cat.cex = 1.2,
  print.mode = c("percent", "raw"),
  cat.default.pos = "outer",
  ...
)
}
\arguments{
\item{object}{see \emph{Usage}}

\item{palette}{\link[base]{character} scalar, color palette to use.
Default \code{'ggplot'} uses the default color palette as seen in package \CRANpkg{ggplot2},
with workhorse function \link[scales]{pal_hue}.
Other options are '\code{rainbow}' with workhorse function \link[grDevices]{rainbow}.}

\item{ind, lty, fill, alpha, cex, cat.col, cat.cex, cat.fontface, print.mode, cat.default.pos, ...}{see function \link[VennDiagram]{draw.pairwise.venn}}
}
\value{
Function \link{venn} returns an S3 object \code{'venn'}, which inherits from class \link[grid]{gList}.
}
\description{
Minor modifications to function \link[VennDiagram]{venn.diagram}.
}
\details{
Workhorse of function \link{venn} is one of
\link[VennDiagram]{draw.single.venn},
\link[VennDiagram]{draw.pairwise.venn},
\link[VennDiagram]{draw.triple.venn},
\link[VennDiagram]{draw.quad.venn},
or
\link[VennDiagram]{draw.quintuple.venn}.
}
\examples{
venn(list(
  A = state.name[1:20], 
  B = state.name[2:21], 
  C = state.name[3:22],
  D = state.name[4:23])) # more friendly to naked eye
venn(list(
  A = state.name[1:20], 
  B = state.name[2:21], 
  C = state.name[3:22],
  D = state.name[4:23]), palette = 'rainbow')
}
\seealso{
\link[VennDiagram]{venn.diagram} does handle \link[base]{list} input, but not as elegantly as
function \link{venn.matrix}.
}
