% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprintf.venn.R
\name{Sprintf.venn}
\alias{Sprintf.venn}
\title{Sprintf.venn}
\usage{
Sprintf.venn(model, ...)
}
\arguments{
\item{model}{a \link{venn} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{Sprintf.venn} returns a \link[base]{character} scalar.
}
\description{
Sprintf.venn
}
