% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{venn}
\alias{venn}
\alias{venn.list}
\alias{venn.data.frame}
\alias{venn.matrix}
\title{Venn Diagram via \CRANpkg{VennDiagram}}
\usage{
venn(object, ...)

\method{venn}{list}(object, ...)

\method{venn}{data.frame}(object, ...)

\method{venn}{matrix}(
  object,
  ind = FALSE,
  lty = "blank",
  fill = rainbow(n = n_cat),
  alpha = 0.25,
  cex = 1,
  cat.col = fill,
  cat.fontface = "bold",
  cat.cex = 1.2,
  print.mode = c("percent", "raw"),
  cat.default.pos = "outer",
  ...
)
}
\arguments{
\item{object}{..}

\item{...}{additional parameters of \link[VennDiagram]{draw.pairwise.venn}, etc.}

\item{ind}{\link[base]{logical} scalar (see \link[VennDiagram]{draw.pairwise.venn}),
whether the function is to automatically draw the diagram before returning the \link[grid]{gList} object or not}

\item{lty}{\link[base]{character} scalar (see \link[VennDiagram]{draw.pairwise.venn}),
line dash pattern of the circles' circumferences.
Default \code{'blank'} disables circle circumference.}

\item{fill}{\link[base]{character} \link[base]{vector} (see \link[VennDiagram]{draw.pairwise.venn}), colors of the circles' areas.
Default \link[grDevices]{rainbow}}

\item{alpha}{\link[base]{numeric} scalar (see \link[VennDiagram]{draw.pairwise.venn}), transparency of the circles' areas}

\item{cex}{\link[base]{numeric} scalar (see \link[VennDiagram]{draw.pairwise.venn})}

\item{cat.col}{\link[base]{character} \link[base]{vector} (see \link[VennDiagram]{draw.pairwise.venn}), colors of the category names.
Default value being the same as \code{fill}}

\item{cat.fontface}{\link[base]{character} scalar (see \link[VennDiagram]{draw.pairwise.venn}),
font face of the category names.  Default \code{'bold'}}

\item{cat.cex}{\link[base]{numeric} scalar (see \link[VennDiagram]{draw.pairwise.venn}),
size of the category names}

\item{print.mode}{\link[base]{character} scalar or \link[base]{vector} (see \link[VennDiagram]{draw.pairwise.venn}),
Default \code{c('percent', 'raw')}}

\item{cat.default.pos}{..}
}
\value{
Function \link{venn} returns a \link[grid]{gList} object.
}
\description{
For students, encourage them to use \link[VennDiagram]{venn.diagram}.
}
\details{
Workhorse of function \link{venn} is one of
\link[VennDiagram]{draw.single.venn},
\link[VennDiagram]{draw.pairwise.venn},
\link[VennDiagram]{draw.triple.venn},
\link[VennDiagram]{draw.quad.venn},
or
\link[VennDiagram]{draw.quintuple.venn}.

Labels of subsets with zero counts are suppressed and printed as \code{'-'}).
}
\examples{
plot(venn(list(
  A = state.name[1:30], 
  B = state.name[20:45], 
  C = state.name[c(15:40, 46:50)])))

plot(venn(list(
  A = state.name[1:20], 
  B = state.name[2:21], 
  C = state.name[3:22],
  D = state.name[4:23])))

}
\seealso{
\link[VennDiagram]{venn.diagram} does handle \link[base]{list} input, but not as elegantly as
function \link{venn.matrix}.
}
